+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # v0.0.4 (TBD)
-* Feature: Crop
 * Translations (Italian / English)
+
+[![](https://img.shields.io/badge/donate-paypal-005EA6.svg)](https://www.paypal.me/ptkdev) [![](https://img.shields.io/badge/donate-patreon-F87668.svg)](https://www.patreon.com/ptkdev) [![](https://img.shields.io/badge/donate-sponsors-ea4aaa.svg)](https://github.com/sponsors/ptkdev/)  [![](https://img.shields.io/badge/donate-ko--fi-29abe0.svg)](https://ko-fi.com/ptkdev)
+
 
 # v0.0.3 (January 08, 2019)
 * Developer mode: build system
 * Support: Crunchyroll
 * Support: Twitch
+* Feature: Crop
 * Feature: Donate tab
 * Feature: Support tab
 * Feature: Customize Aspect ratio value