+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# v0.0.4 (TBD)
-* Translations (Italian / English)
+# v1.0.0 (TBD)
+* Fix: youtube on/off switch
+* Fix: twitch on/off switch
+* Translations (Italian / English / Polish)
 
 [![](https://img.shields.io/badge/donate-paypal-005EA6.svg)](https://www.paypal.me/ptkdev) [![](https://img.shields.io/badge/donate-patreon-F87668.svg)](https://www.patreon.com/ptkdev) [![](https://img.shields.io/badge/donate-sponsors-ea4aaa.svg)](https://github.com/sponsors/ptkdev/)  [![](https://img.shields.io/badge/donate-ko--fi-29abe0.svg)](https://ko-fi.com/ptkdev)